(Major): Move all parameters back to top scope (nginx vs nginx::config)
