+++ manifests/resource/map.pp
@@ -80,8 +80,8 @@ define nginx::resource::map (
     "Invalid ensure value '${ensure}'. Expected 'present' or 'absent'")
   if ($default != undef) { validate_string($default) }
 
-  $root_group = $::nginx::config::root_group
-  $conf_dir   = "${::nginx::config::conf_dir}/conf.d"
+  $root_group = $::nginx::root_group
+  $conf_dir   = "${::nginx::conf_dir}/conf.d"
 
   $ensure_real = $ensure ? {
     'absent' => absent,
@@ -94,7 +94,7 @@ define nginx::resource::map (
     mode  => '0644',
   }
 
-  file { "${conf_dir}/${name}-map.conf":
+  file { "${::nginx::conf_dir}/conf.d/${name}-map.conf":
     ensure  => $ensure_real,
     content => template('nginx/conf.d/map.erb'),
     notify  => Class['::nginx::service'],
