+++ manifests/resource/upstream/member.pp
@@ -59,7 +59,7 @@ define nginx::resource::upstream::member (
 
   # Uses: $server, $port, $upstream_fail_timeout
   concat::fragment { "${upstream}_upstream_member_${name}":
-    target  => "${::nginx::config::conf_dir}/conf.d/${upstream}-upstream.conf",
+    target  => "${::nginx::conf_dir}/conf.d/${upstream}-upstream.conf",
     order   => 40,
     content => template('nginx/conf.d/upstream_member.erb'),
   }
