+++ spec/defines/resource_mailhost_spec.rb
@@ -9,12 +9,12 @@ describe 'nginx::resource::mailhost' do
       ipv6_enable: true
     }
   end
-  let(:pre_condition) { ['include ::nginx::config'] }
+  let(:pre_condition) { ['include ::nginx'] }
 
   describe 'os-independent items' do
     describe 'basic assumptions' do
       let(:params) { default_params }
-      it { is_expected.to contain_class('nginx::config') }
+      it { is_expected.to contain_class('nginx') }
       it { is_expected.to contain_concat("/etc/nginx/conf.mail.d/#{title}.conf").that_requires('File[/etc/nginx/conf.mail.d]') }
       it do
         is_expected.to contain_concat("/etc/nginx/conf.mail.d/#{title}.conf").with('owner' => 'root',
