+++ spec/defines/resource_vhost_spec.rb
@@ -4,6 +4,7 @@ describe 'nginx::resource::vhost' do
   let :title do
     'www.rspec.example.com'
   end
+
   let :default_params do
     {
       www_root: '/',
@@ -11,21 +12,23 @@ describe 'nginx::resource::vhost' do
       listen_unix_socket_enable: true
     }
   end
+
   let :facts do
     {
       ipaddress6: '::'
     }
   end
+
   let :pre_condition do
     [
-      'include ::nginx::config'
+      'include ::nginx'
     ]
   end
 
   describe 'os-independent items' do
     describe 'basic assumptions' do
       let(:params) { default_params }
-      it { is_expected.to contain_class('nginx::config') }
+      it { is_expected.to contain_class('nginx') }
       it do
         is_expected.to contain_concat("/etc/nginx/sites-available/#{title}.conf").with('owner' => 'root',
                                                                                        'group' => 'root',
@@ -46,7 +49,7 @@ describe 'nginx::resource::vhost' do
     describe 'with $confd_only enabled' do
       let(:pre_condition) { 'class { "nginx": confd_only => true }' }
       let(:params) { default_params }
-      it { is_expected.to contain_class('nginx::config') }
+      it { is_expected.to contain_class('nginx') }
       it do
         is_expected.to contain_concat("/etc/nginx/conf.d/#{title}.conf").with('owner' => 'root',
                                                                               'group' => 'root',
